# optimize LayeredMap::get()