--- a/experimental/storage/layered-map/src/lib.rs
+++ b/experimental/storage/layered-map/src/lib.rs
@@ -7,6 +7,7 @@ pub use map::LayeredMap;
 use std::hash::Hash;
 
 mod dropper;
+mod flatten_perfect_tree;
 pub mod iterator;
 mod layer;
 mod map;
@@ -15,6 +16,7 @@ mod node;
 pub(crate) mod r#ref;
 #[cfg(test)]
 mod tests;
+mod utils;
 
 /// When recursively creating a new `MapLayer` (a crit bit tree overlay), partitioning and passing
 /// down `Vec<(K, Option<V>)>` would mean a lot of memory allocation. That's why we require