## Description
<!-- Please include a summary of the change, including which issue it fixes or what feature it adds. Include relevant motivation, context and documentation as appropriate. List dependencies that are required for this change, if any. -->

Compress the top of the critbit tree by using a newly introduced FlattenedPerfectTree, which omits the internal nodes.

<details>
<summary>
Due to the complexity of the code, update is slower.
</summary>

```
insert_in_batches/layered_map_1_batches_of_10k_updates
                        time:   [1.6578 ms 1.6878 ms 1.7120 ms]
                        thrpt:  [5.9814 Melem/s 6.0671 Melem/s 6.1768 Melem/s]
                 change:
                        time:   [+14.571% +18.106% +21.934%] (p = 0.00 < 0.05)
                        thrpt:  [-17.988% -15.331% -12.718%]
                Performance has regressed.
insert_in_batches/layered_map_8_batches_of_10k_updates
                        time:   [21.700 ms 21.854 ms 22.016 ms]
                        thrpt:  [3.7210 Melem/s 3.7485 Melem/s 3.7752 Melem/s]
                 change:
                        time:   [-2.5688% -1.7149% -0.9398%] (p = 0.00 < 0.05)
                        thrpt:  [+0.9487% +1.7449% +2.6365%]
                        Change within noise threshold.
insert_in_batches/layered_map_1_batches_of_100k_updates                                                                                                                                                                                                             time:   [18.788 ms 18.974 ms 19.140 ms]
                        thrpt:  [5.3500 Melem/s 5.3970 Melem/s 5.4503 Melem/s]
                 change:
                        time:   [+25.470% +26.981% +28.499%] (p = 0.00 < 0.05)
                        thrpt:  [-22.178% -21.248% -20.300%]
                        Performance has regressed.
Benchmarking insert_in_batches/layered_map_8_batches_of_100k_updates: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 41.8s, or reduce sample count to 10.
insert_in_batches/layered_map_8_batches_of_100k_updates
                        time:   [405.39 ms 406.99 ms 408.59 ms]
                        thrpt:  [2.0050 Melem/s 2.0128 Melem/s 2.0208 Melem/s]
                 change:
                        time:   [+11.535% +11.973% +12.406%] (p = 0.00 < 0.05)
                        thrpt:  [-11.037% -10.693% -10.342%]
                        Performance has regressed.
```
</details>


<details>
<summary>
However reading is greatly faster
</summary>

```
get_existing/layered_map_1000k_items
                        time:   [1.0122 ms 1.0130 ms 1.0139 ms]
                        thrpt:  [10.099 Melem/s 10.109 Melem/s 10.117 Melem/s]
                 change:
                        time:   [-63.890% -63.330% -62.865%] (p = 0.00 < 0.05)
                        thrpt:  [+169.29% +172.70% +176.93%]
                        Performance has improved.
Benchmarking get_existing/layered_map_128000k_items: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 7.4s, enable flat sampling, or reduce sample count to 50.
get_existing/layered_map_128000k_items
                        time:   [1.3693 ms 1.3753 ms 1.3824 ms]
                        thrpt:  [7.4073 Melem/s 7.4457 Melem/s 7.4783 Melem/s]
                 change:
                        time:   [-78.212% -78.015% -77.792%] (p = 0.00 < 0.05)
                        thrpt:  [+350.30% +354.86% +358.98%]
                        Performance has improved.
```

</details>

<details>
<summary>
Net-net, Hexy update is significantly faster
</summary>

``` text                                                                                                            
Benchmarking hexy_updates/hexy_update_leaves_128m_batch_10k_pipeline_depth_0: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 6.1s, or reduce sample count to 80.
hexy_updates/hexy_update_leaves_128m_batch_10k_pipeline_depth_0
                        time:   [61.543 ms 61.747 ms 61.953 ms]
                        thrpt:  [165.29 Kelem/s 165.84 Kelem/s 166.39 Kelem/s]
                 change:
                        time:   [-3.0909% -2.6292% -2.1623%] (p = 0.00 < 0.05)
                        thrpt:  [+2.2101% +2.7002% +3.1895%]
                        Performance has improved.
Benchmarking hexy_updates/hexy_update_leaves_128m_batch_10k_pipeline_depth_2: Warming up for 3.0000 s                                                                                                                                       Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 9.9s, or reduce sample count to 50.
hexy_updates/hexy_update_leaves_128m_batch_10k_pipeline_depth_2
                        time:   [102.99 ms 103.91 ms 104.86 ms]
                        thrpt:  [97.653 Kelem/s 98.548 Kelem/s 99.423 Kelem/s]
                 change:
                        time:   [-28.540% -27.875% -27.126%] (p = 0.00 < 0.05)
                        thrpt:  [+37.223% +38.647% +39.938%]
                        Performance has improved.
Benchmarking hexy_updates/hexy_update_leaves_128m_batch_10k_pipeline_depth_8: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 16.0s, or reduce sample count to 30.
hexy_updates/hexy_update_leaves_128m_batch_10k_pipeline_depth_8
                        time:   [162.87 ms 163.54 ms 164.27 ms]
                        thrpt:  [62.337 Kelem/s 62.615 Kelem/s 62.872 Kelem/s]
                 change:
                        time:   [-27.008% -26.279% -25.573%] (p = 0.00 < 0.05) 
                        thrpt:  [+34.360% +35.647% +37.000%]
                        Performance has improved. 
```
</details>

## Type of Change
- [x] Performance improvement


## Which Components or Systems Does This Change Impact?
- [x] Validator Node


## How Has This Been Tested?
<!--
- Please ensure that the functionality introduced by this change is well tested and verified to work as expected.
- Ensure tests cover both happy and unhappy paths.
- List and link relevant tests.
-->

unit tests
benchmarks